Delsys-EMG/requirements.txt
PySide6
Vispy
Scipy
pythonnet
numpy
paho.mqtt.client 
time
socket
pandas
csv
json